@@ -7,59 +7,66 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "postinstall": "./node_modules/.bin/webpack",
+    "postinstall": "webpack",
     "start": "node src/app.js",
-    "test": "./node_modules/.bin/mocha test/**/*.js"
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "babel-core": "^6.25.0",
-    "babel-loader": "^7.1.1",
-    "babel-plugin-transform-es2015-modules-commonjs": "^6.24.1",
+    "babel-cli": "^6.26.0",
+    "babel-core": "^6.26.0",
+    "babel-eslint": "^8.0.1",
+    "babel-jest": "^21.2.0",
+    "babel-loader": "^7.1.2",
+    "babel-plugin-transform-es2015-modules-commonjs": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-es2016": "^6.24.1",
     "babel-preset-es2017": "^6.24.1",
+    "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "^6.24.1",
     "chai": "^3.5.0",
     "css-loader": "^0.26.4",
     "eslint": "^3.19.0",
+    "eslint-plugin-flowtype": "^2.39.1",
     "eslint-plugin-react": "^6.10.3",
-    "extract-text-webpack-plugin": "^2.1.2",
+    "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^0.11.2",
+    "flow-bin": "^0.57.3",
     "isparta": "^4.0.0",
-    "mocha": "^3.4.2",
+    "jest": "^21.2.1",
+    "mocha": "^3.5.3",
+    "react-test-renderer": "^16.0.0",
     "sinon": "^1.17.7",
     "style-loader": "^0.13.2",
-    "webpack": "^3.3.0",
-    "webpack-dev-server": "^2.5.1"
+    "webpack": "^3.8.1",
+    "webpack-dev-server": "^2.9.3"
   },
   "dependencies": {
-    "body-parser": "^1.17.2",
-    "connect-redis": "^3.3.0",
+    "body-parser": "^1.18.2",
+    "connect-redis": "^3.3.2",
     "cookie": "^0.3.1",
     "cookie-parser": "^1.4.3",
     "csurf": "^1.9.0",
-    "csv": "^1.1.1",
+    "csv": "^1.2.1",
     "dotenv": "^4.0.0",
-    "express": "^4.15.3",
-    "express-session": "^1.15.3",
+    "express": "^4.16.2",
+    "express-session": "^1.15.6",
     "firebase": "^4.6.0",
-    "firebase-admin": "^5.4.2",
-    "immutable": "^3.8.1",
-    "jsonwebtoken": "^7.4.1",
-    "mysql": "^2.13.0",
+    "immutable": "^4.0.0-rc.9",
+    "jsonwebtoken": "^7.4.3",
+    "mysql": "^2.15.0",
     "path": "^0.12.7",
-    "pixi.js": "^4.5.5",
-    "pug": "^2.0.0-rc.2",
+    "pixi.js": "^4.6.0",
+    "pug": "^2.0.0-rc.4",
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "react-redux": "^5.0.6",
-    "redis": "^2.7.1",
+    "redis": "^2.8.0",
     "redux": "^3.7.2",
     "socket.io": "^1.7.4",
-    "socket.io-client": "^2.0.3",
-    "uid-safe": "^2.1.4"
+    "socket.io-client": "^2.0.4",
+    "uid-safe": "^2.1.5"
   }
 }
