@@ -1,23 +1,30 @@
 import { Record } from 'immutable';
 
 import Room from '../models/Room.js';
-import Game from '../models/Game.js';
 import UI from './UI.js';
 
 export default class State extends Record({
   userId: null,
+  deck: null,
   room: null,
   me: null,
   opponent: null,
   ui: new UI(),
 }) {
 
-  init(userId) {
-    return this.set('userId', userId);
+  init({ userId, deck }) {
+    return this.withMutations(mnt => {
+      mnt.set('userId', userId)
+        .set('deck', deck);
+    });
   }
 
   syncRoom(data) {
     const room = Room.restore(data);
+    return this.setRoom(room);
+  }
+
+  setRoom(room) {
     return this.withMutations(mnt => {
       mnt.set('room', room)
         .set('me', room.player(mnt.userId))
@@ -34,7 +41,7 @@ export default class State extends Record({
 
   syncGame(data) {
     return this.withMutations(mnt => {
-      mnt.set('room', mnt.room.set('game', Game.restore(data)))
+      mnt.set('room', mnt.room.syncGame(data))
         .set('ui', mnt.ui.clear());
     });
   }
@@ -73,24 +80,24 @@ export default class State extends Record({
     }
     const result = {
       me: {
-        name: unit.status().name,
+        name: unit.status.name,
         hp: unit.hp,
         offense: unit.offense
       },
       tg: {
-        name: target.status().name,
+        name: target.status.name,
         hp: target.hp,
         offense: target.offense
       }
     };
-    if (unit.klass().healer) {
-      result.me.val = unit.status().pow;
+    if (unit.klass.healer) {
+      result.me.val = unit.status.pow;
     } else {
       result.me.val = target.effectValueBy(unit);
       result.me.hit = target.hitRateBy(unit, game.field.avoidance(cellId));
       result.me.crit = target.critRateBy(unit);
     }
-    if (!unit.klass().healer && !target.klass().healer) {
+    if (!unit.klass.healer && !target.klass.healer) {
       // counter attack
       if (game.checkActionable(target, cellId, ui.movedCell)) {
         result.tg.val = unit.effectValueBy(target);
@@ -161,9 +168,17 @@ export default class State extends Record({
     if (this.canAct(cellId)) {
       const actCell = (cellId != ui.movedCell) ? cellId : undefined;
       setImmediate(() => {
-        onAct(ui.forcusedCell, ui.movedCell, actCell);
+        if (onAct && typeof onAct === 'function') {
+          onAct(ui.forcusedCell, ui.movedCell, actCell);
+        }
       });
-      return this.set('ui', ui.act());
+      return this.withMutations(mnt => {
+        mnt.set('ui', ui.act());
+        if (mnt.room.isSolo) {
+          mnt.set('room', mnt.room.actInGame(mnt.userId, ui.forcusedCell, ui.movedCell, actCell))
+            .clearUI();
+        }
+      });
     }
     return this.undo().clearUI();
   }
@@ -174,4 +189,106 @@ export default class State extends Record({
     return cellId == ui.movedCell
       || game.checkActionable(ui.forcusedUnit, ui.movedCell, cellId);
   }
+
+
+
+  isSolo() {
+    return this.room != null && this.room.isSolo;
+  }
+
+  isCOMTurn() {
+    return !this.room.isTrunPlayer(this.userId);
+  }
+
+  startSoloPlay() {
+    return this.setRoom(Room.soloRoom(this.userId, this.deck));
+  }
+
+  selectUnitSolo(list) {
+    return this.setRoom(
+      this.room.selectUnits(this.userId, list).mightEngage()
+    );
+  }
+
+  endTurnSolo() {
+    return this.set('room', this.room.endTurn(this.userId));
+  }
+
+  mightStartAITurn() {
+    if (!this.isCOMTurn()) {
+      return this;
+    }
+    return this.withMutations(mnt => {
+      for (let i=0; i<20; i++) {
+        if (!mnt.isCOMTurn() || mnt.room.game.isEnd) {
+          break;
+        }
+        mnt.actByAI();
+      }
+    });
+  }
+
+  actByAI() {
+    const { room, ui } = this;
+    const com = room.opponent(this.userId);
+    const { game } = room;
+    const units = game.ownedUnits(com.offense).filter(unit => !unit.acted);
+
+    let priUnit, priAction;
+    units.forEach(unit => {
+      const bufferUi = ui.forcus(unit).setRange(game);
+      if (!unit.klass.healer) {
+        // 行動対象セル抽出
+        const targetCellIds = bufferUi.ranges.getActionables().filter(acell => {
+          const target = game.unit(acell);
+          return target && target.offense !== unit.offense;
+        });
+        // 行動パラメータを算出
+        const actions = targetCellIds.map(tcell => {
+          const froms = bufferUi.ranges.getActionableFrom(tcell);
+          let from;
+          // 空のfromセルを抽出
+          froms.forEach(_from => {
+            if (game.unit(_from) == null) {
+              from = _from;
+            }
+          });
+          const target = game.unit(tcell);
+          return {
+            from,
+            to: tcell,
+            value: (
+              // 行動評価値
+              from != undefined
+              ? target.expectedEvaluationBy(unit, game.field.avoidance(target.cellId))
+              : null
+            )
+          };
+        }).filter(actionCell => {
+          return actionCell.from != undefined;
+        });
+
+        if (actions.length > 0) {
+          // 最大評価値の行動抽出
+          const action = actions.reduce((pre, cur) => {
+            return pre.value < cur.value ? cur : pre;
+          });
+          if (!priAction || priAction.value < action.value) {
+            priAction = action;
+            priUnit = unit;
+          }
+        }
+      }
+    });
+    if (priUnit && priAction) {
+      // 行動確定
+      return this.set(
+        'room',
+        this.room.actInGame(com.id, priUnit.cellId, priAction.from, priAction.to)
+      );
+    } else {
+      return this.set('room', this.room.endTurn(com.id));
+    }
+  }
+
 }
