@@ -1,10 +1,23 @@
+// @flow
 import React from 'react';
 import './Selector.css';
 
 import UnitMaster from '../../../data/json/unit.json';
 
-export default class Selector extends React.Component {
-  constructor(props) {
+type Props = {
+  costLimit: number,
+  myDeck: Array<number>,
+  isOffense: boolean,
+  opponentsDeck: Array<number>,
+  onSubmit: Array<number> => void,
+};
+type State = {
+  selected: Array<any>,
+  isEmitted: boolean,
+};
+
+export default class Selector extends React.Component<Props, State> {
+  constructor(props: Props) {
     super(props);
     this.state = {
       selected: [],
@@ -12,7 +25,7 @@ export default class Selector extends React.Component {
     };
   }
 
-  selectSortie(index) {
+  selectSortie(index: number) {
     const indexOf = this.state.selected.indexOf(index);
     if (indexOf >= 0) {
       this.setState({
@@ -28,18 +41,15 @@ export default class Selector extends React.Component {
   }
 
   render() {
-    const { me, opponent, onSubmit } = this.props;
-    if (!me || !opponent) {
-      return null;
-    }
-    let { costLimit=1 } = this.props;
-    if (me.offense == false) {
+    const { myDeck, isOffense, opponentsDeck, onSubmit } = this.props;
+    let { costLimit } = this.props;
+    if (isOffense === false) {
       costLimit += 2;
     }
     const { selected } = this.state;
 
-    const army = me.deck.map(id => UnitMaster[id]);
-    const enemy = opponent.deck.map(id => UnitMaster[id]);
+    const army = myDeck.map(id => UnitMaster[id]);
+    const enemy = opponentsDeck.map(id => UnitMaster[id]);
 
     let cost = 0;
     selected.forEach(i => {
@@ -58,12 +68,12 @@ export default class Selector extends React.Component {
         <div id="sel-container">
           <div className="sel-row">
             <div className="sel-side">自軍</div>
-            <div className={'sel-box ' + (me.offense ? 'box-offense' : 'box-defense')}>
-              <div className="sel-roll">{me.offense ? '攻撃' : '防衛'}</div>
+            <div className={'sel-box ' + (isOffense ? 'box-offense' : 'box-defense')}>
+              <div className="sel-roll">{isOffense ? '攻撃' : '防衛'}</div>
               <ul className="sel-units-list">
                 {army.map((unit, i) => {
                   const isSelected = (selected.indexOf(i) >= 0);
-                  const imageName = `${unit.klass}_${!isSelected ? 0 : me.offense ? 1 : 2}`;
+                  const imageName = `${unit.klass}_${!isSelected ? 0 : isOffense ? 1 : 2}`;
                   return (
                     <li
                       key={i}
@@ -72,7 +82,7 @@ export default class Selector extends React.Component {
                         if (this.state.isEmitted) {
                           return;
                         }
-                        this.selectSortie(i, army[i].cost);
+                        this.selectSortie(i);
                       }}>
                       <img src={`/image/units/${imageName}.png`} />
                       <div className="sel-unit-name">{unit.name}</div>
@@ -104,8 +114,8 @@ export default class Selector extends React.Component {
           </div>
           <div className="sel-row">
             <div className="sel-side">敵軍</div>
-            <div className={'sel-box ' + (opponent.offense ? 'box-offense' : 'box-defense')}>
-              <div className="sel-roll">{opponent.offense ? '攻撃' : '防衛'}</div>
+            <div className={'sel-box ' + (!isOffense ? 'box-offense' : 'box-defense')}>
+              <div className="sel-roll">{!isOffense ? '攻撃' : '防衛'}</div>
               <ul className="sel-units-list">
                 {enemy.map((unit, i) => {
                   return (
